Finished APIs

+++ src/main/java/Controllers/CatererController.java
@@ -1,78 +1,123 @@
 package Controllers;
 
 import Server.Main;
+import org.glassfish.jersey.media.multipart.FormDataParam;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
 
+import javax.ws.rs.Consumes;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 
+@Path("caterer/")
 public class CatererController {
 
-    public static void insertCaterer(String catererName, String foodType, Float priceHead) {
+    @POST
+    @Path("insert")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String insertCaterer(@FormDataParam("name") String catererName, @FormDataParam("type") String foodType,
+                                    @FormDataParam("price") Float priceHead) {
 
         try {
+            if (catererName == null || foodType == null || priceHead == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("caterer/new name=" + catererName);
             PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Caterers(CatererName, FoodType, PriceHead) VALUES (?, ?, ?)");
 
             ps.setString(1, catererName);
             ps.setString(2, foodType);
             ps.setFloat(3, priceHead);
             ps.executeUpdate();
-            System.out.println("Record added to Caterers table");
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
         }
 
     }
 
-    public static void deleteCaterer(int catererID) {
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String deleteCaterer(@FormDataParam("id") Integer catererID) {
+
         try{
+            if (catererID == null) {
+                throw new Exception("ID is missing in the HTTP request.");
+            }
+            System.out.println("caterer/delete id=" + catererID);
 
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Caterers WHERE CatererID = ?");
             ps.setInt(1, catererID);
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception e) {
-            System.out.println(e.getMessage());
+            System.out.println("Database error: " + e.getMessage());
+            return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
         }
     }
 
-    public static void updateCaterer(String catererName, String foodType, Float priceHead) {
+    @POST
+    @Path("update")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String updateCaterer(@FormDataParam("id") Integer catererID, @FormDataParam("name") String catererName, @FormDataParam("type") String foodType,
+                                    @FormDataParam("price") Float priceHead) {
 
         try {
+            if (catererID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("entertainment/update id=" + catererID);
+
             PreparedStatement ps = Main.db.prepareStatement("UPDATE Caterers SET CatererName = ?, FoodType = ?, PriceHead = ? WHERE CatererID = ?");
 
             ps.setString(1, catererName);
             ps.setString(2, foodType);
             ps.setFloat(3, priceHead);
-
+            ps.setInt(4, catererID);
             ps.execute();
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
         }
     }
 
-    public static void listCaterer() {
-
+    @POST
+    @Path("list")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String listCaterer() {
+        System.out.println("caterer/list");
+        JSONArray list = new JSONArray();
         try {
 
             PreparedStatement ps = Main.db.prepareStatement("SELECT CatererID, CatererName, FoodType, PriceHead FROM Caterers");
 
             ResultSet results = ps.executeQuery();
             while (results.next()) {
-                int catererID = results.getInt(1);
-                String catererName = results.getString(2);
-                String foodType = results.getString(3);
-                Float priceHead = results.getFloat(4);
-                System.out.print("catererID: " + catererID + ", ");
-                System.out.print("Caterer Name: " + catererName + ", ");
-                System.out.print("Type: " + foodType + ", ");
-                System.out.print("Price/head: " + priceHead);
-                System.out.println(" ");
+                JSONObject item = new JSONObject();
+                item.put("id", results.getInt(1));
+                item.put("name", results.getString(2));
+                item.put("type", results.getString(3));
+                item.put("price", results.getFloat(4));
+                list.add(item);
             }
-
+            return list.toString();
         } catch (Exception exception) {
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
         }
     }
 }
